refactor(application): promote test_helpers to pub(crate) and relocate metadata_builder tests
- Extract test_helpers from nested mod tests into a top-level
  #[cfg(test)] pub(crate) mod test_helpers in mod.rs
- Change all pub(super) visibility to pub(crate) for both the module
  and its 7 helper functions
- Relocate 3 metadata_builder unit tests from mod.rs to
  metadata_builder.rs with #[cfg(test)] mod tests
- Update use test_helpers as th to use super::test_helpers as th
  in mod.rs tests block

Establishes the pub(crate) sharing pattern for sibling builder modules.
All 39 existing tests continue to pass.

Closes #531
